Change website structure and add header on homepage
Since the website is being redesigned, I changed some parts of the the way the website's code structure is organized. This mostly includes updating the default template as well as adding in a feature to inject scripts, stylesheets, and Google fonts directly from a web page's YAML front matter.

Additionally, I updated the header to match the Figma designs, and added the full header on the homepage with the YouTube video and welcome message.