Converted site to using Webpack
I added the jekyll-webpack plugin and converted the structure of the site to using Webpack. This means TypeScript and SCSS files are converted via Webpack now instead of through Jekyll. This uses GitHub actions to build the website from scratch on GitHub and then deploy to the gh-pages branch, which is where the website is hosted.